requirements.txt
# 3D Teeth Segmentation PyTorch Requirements
# Core PyTorch packages
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# 3D processing
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0
torch-cluster>=1.6.0

# Point cloud and mesh processing
trimesh>=3.20.0
open3d>=0.17.0
pytorch3d>=0.7.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
plotly>=5.0.0
seaborn>=0.11.0
pandas>=1.3.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Training utilities
tensorboard>=2.8.0
wandb>=0.13.0
tqdm>=4.60.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.2.0
hydra-core>=1.2.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Development
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0